Adding repository and VM

+++ app/src/main/java/com/example/android/data/dto/ArticlesResultDto.kt
@@ -3,7 +3,7 @@ package com.example.android.data.dto
 data class ArticlesResultDto(
     val uri: String?,
     val url: String?,
-    val id: Int?,
-    val asset_id: Int?,
+    val id: Double?,
+    val asset_id: Double?,
     val source: String?
 )
